Backend/requirements.txt
aioboto3==15.5.0
boto3==1.40.61
kubernetes==32.0.1
uvicorn==0.34.2
fastapi==0.115.12
websockets==10.1
email_validator==2.2.0
PyJWT==2.10.1
